genre 删除校验：存在book引用该genre时无法删除

+++ src/service/genreService.js
@@ -1,9 +1,10 @@
 const Genre = require('../models/genre')
+const BookService = require('./bookService')
 
 exports.create = req => {
   return new Promise((resolve, reject) => {
     const name = req.body.name
-    Genre.findOne({ name: name }, (err, result) => {
+    Genre.findOne({name: name}, (err, result) => {
       if (err) {
         return reject(err)
       }
@@ -33,12 +34,17 @@ exports.delete = id => {
       if (result == null) {
         return reject('找不到该类型')
       }
-      Genre.findByIdAndRemove(id, err => {
-        if (err) {
-          return reject(err)
+      BookService.count({genre: result}).then(count => {
+        if (count > 0) {
+          return reject('库中存在该类型书籍，无法删除')
         }
-        return resolve({ id })
-      })
+        Genre.findByIdAndRemove(id, err => {
+          if (err) {
+            return reject(err)
+          }
+          return resolve({id})
+        })
+      }).catch(err => reject(err))
     })
   })
 }
@@ -47,15 +53,14 @@ exports.update = req => {
   return new Promise((resolve, reject) => {
     const _id = req.body._id
     const name = req.body.name
-    // TODO 不可修改为已存在的类型
-    Genre.findOne({ name }, function (err, genre) {
+    Genre.findOne({name}, function (err, genre) {
       if (err) {
         return reject(err)
       }
       if (genre) {
         return reject('已存在该类型-' + name)
       }
-      Genre.updateOne({ _id }, { name }, {}, function (err, result) {
+      Genre.updateOne({_id}, {name}, {}, function (err, result) {
         if (err) {
           return reject(err)
         }
@@ -81,14 +86,14 @@ exports.detail = id => {
 
 exports.list = query => {
   return new Promise((resolve, reject) => {
-    let { name, page, limit } = query
+    let {name, page, limit} = query
     console.log(query)
     page = page ? page : 1
     limit = limit ? limit : 10
     let query_ = {}
     let like = []
     if (name) {
-      like.push({ name: { $regex: new RegExp(name, 'i') } })
+      like.push({name: {$regex: new RegExp(name, 'i')}})
       query_.$or = like
     }
     Genre.find(query_, '-__v')
@@ -102,7 +107,7 @@ exports.list = query => {
           if (err) {
             return reject(err)
           }
-          resolve({ items: result, total: count })
+          resolve({items: result, total: count})
         })
       })
   })
